[FLINK-12966]finalize package name of Hive table source/sink
